pyproject.toml
[project]
name = "django-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<4"
dependencies = [
    "django>=5.2",
    "django-environ>=0.12.0,<4",
    "flake8>=7.2.0",
    "psycopg2>=2.9.10",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "pre-commit (>=4.2.0,<5.0.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "drf-spectacular[pydantic] (>=0.28.0,<0.29.0)",
    "punq (>=0.7.0,<0.8.0)",
    "python-dateutil (>=2.9.0.post0,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
]
[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
pre-commit = "^4.2.0"
black = "^25.1.0"


[tool.black]
line-length = 120
target-version = ["py311"]
extend-exclude = '(venv|alembic|docker|\.git|__pycache__)'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
line_length = 120
lines_after_imports = 2
lines_between_sections = 1
lines_between_types = 0
force_alphabetical_sort_within_sections = true
ensure_newline_before_comments = true
combine_as_imports = true
combine_star = true
skip_gitignore = true
balanced_wrapping = true
force_grid_wrap = 2
src_paths = ["apps", "common", "tests"]
skip_glob = [".git", ".idea", ".vscode", "venv"]
known_fastapi = ["fastapi"]
known_sqlalchemy = ["sqlalchemy"]
known_first_party = ["core"]
sections = ["FUTURE", "STDLIB", "FASTAPI", "SQLALCHEMY", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
